casting
git-svn-id: https://svn.ego-gw.it/svn/advsw/Omicron/trunk@56934 47e613b6-93f3-4244-b775-e06b1756610f

+++ src/Odata.cc
@@ -11,9 +11,9 @@ Odata::Odata(const int aChunkDuration, const int aSegmentDuration,
 ////////////////////////////////////////////////////////////////////////////////////
  
   // save parameters
-  ChunkDuration      = fabs(aChunkDuration);
-  SegmentDuration    = fabs(aSegmentDuration);
-  OverlapDuration    = fabs(aOverlapDuration);
+  ChunkDuration      = (int)fabs(aChunkDuration);
+  SegmentDuration    = (int)fabs(aSegmentDuration);
+  OverlapDuration    = (int)fabs(aOverlapDuration);
   fVerbosity         = aVerbosity;
   status_OK=true;
